temporarily use ifdef to separate intel/lavapipe; lavapipe version compiles

+++ libcuda/cuda_api_object.h
@@ -16,7 +16,7 @@
 typedef std::list<gpgpu_ptx_sim_arg> gpgpu_ptx_sim_arg_list_t;
 
 #ifndef OPENGL_SUPPORT
-typedef unsigned long GLuint;
+typedef unsigned int GLuint;
 #endif
 
 struct glbmap_entry {
